Add semantic understanding of the NHS DM&D.
The DM&D is the dictionary of medicines and devices. This is published by the HSCIC (see http://www.nhsbsa.nhs.uk/PrescriptionServices/Documents/PrescriptionServices/dmd_Implementation_Guide_Secondary_Care.pdf) to support electronic prescribing. This logic (and the tests associated with it) demonstrate the internal structures of DM&D and how to find, say, the actual medicinal products from a given virtual medicinal product.