Improvements to Lucene search API.
Rather than using a static global index searcher, we encapsulate our SNOMED-CT searching functionality with a singleton Search object. In addition, the Search object can be configured at runtime to create or read an index at an arbitrary location. These changes are also required for switching to the latest Lucene version which is a high priority for the next stages of development.