Add synonym search support.
This allows one to find the synonyms for an entered term. For example, entering “heart attack” will search for descriptions matching this, take all of the concepts relating to these descriptions and then return all of the synonyms for each of those concepts. This is useful in providing more sophisticated search of clinic letters for example, so that a user may search for “MI” and be pleasantly surprised to get documents matching “myocardial infarction” for example.