Add new fluent SNOMED-CT API.
Migrate to a new search request API in which a request is built using a fluent API and then executed. Add new filtering to search requests. Add concept status to lucene index so that we can optionally only return active concepts. Add unit tests to exercise this new API in searching for diagnostic concepts or medication concepts.