Fix broken search.
I had incorrectly replaced the standard analyser with a whitespace analyser, which was case-sensitive. A custom analyser followed but then terms such as amlodipine+atorvastatin were treated as one term and scored highly on a search for “amlodi”. The new search query using a combination of a queryparser and prefix search solves this and so we can return to using the standard analyser. This passes all tests.

The reason previous tests didn’t show up this failure was that index creation was not deleting old documents - see prior commit that fixes this problem.