Improve full-text search to use default tokenizer from lucene.
I can stop using the QueryParser and generate the query purely from the terms entered by using the same tokenizer that I was using for the QueryParser and yet get the benefit of full control over the query. Here we make each and every term mandatory on a prefix search (unless it is a very short term). This reduces false positive finds and yet still gives excellent search results. I have added a unit test to demonstrate a reduction in false-negative finds.