Fixed IS-A SNOMED-CT search filtering.
Fix the index to index the IS-A direct parent relationships rather than using the cached recursive parent concepts. Add unit tests to catch this error.

Slightly modify the filtering API to cache applied filters and then build a boolean query only when we know all of the filters to apply.