Major work cleaning up code.
Reworked API to be more intuitive, particularly with the difference between direct parents and recursive parents. Modified the Cayenne ORM data map to make this difference clear. Added additional semantics for DM&D as well as more thorough unit testing to ensure that derivation of the DM&D product type is working appropriately. From a VTM or VMP, it is now possible to easily get the trade families in one call.