Test prescribing status.
Most products can be prescribed using a VTM or VMP. There are some products with differences in bioavailability, that should be prescribed as a branded product - an AMP, from a particular manufacturer. Here I start testing this logic within the DM&D in preparation for exposing this functionality to an ePrescribing module.