Tidy up API and packages and switch to using optionals.
Within the DM&D, rather than returning null values (e.g. some VMPs will not have a VTM), return an optional instead to document the possibility of a missing value. In a well constructed DM&D, assuming the authors have not made an error, some relationships will be mandatory but for the moment these are also shown as optional.