Use a simpler custom lucene similarity algorithm.
This derives a scoring factor simply on the inverse of the number of terms, giving priority therefore for results with the fewest terms in any particular search. It is unclear as yet whether this truly improves search results over the prior similarity algorithm. This is a change that does affect the index and so any indices built using the old algorithm should be re-built.