Changes to fluent search API.
Rather than a toggle to choose between using a queryparser or the built-in simple parser to parse the query string, create a dedicated searchUsingQueryParser() method. This means that most users don’t need to catch a ParseException as most users do not use QueryParser. Update unit tests for API changes.