Fix broken relationship import code.
It turns out SELECT…in is non-deterministic in returning things in order. Instead, replace with a slower multi-select for now so it is correct. Hadn’t picked up on this error because our production code still using WebObjects-based version for building parent cache. That code can now be deprecated in favour of this.