Support availability and prescribability for DM&D.
For VTM, VMP, AMP and TF, add support for deriving whether the drug should be shown in pick lists by combining VMP availability status and availability of that product. It may be appropriate to allow selection of drugs that are not prescribable when recording historical drugs, but this logic supports limiting or alerting should the client application need it.